feat(rest): add support for external Express routes with OpenAPI spec
Refactor StaticAssetsRoute to support arbitrary Express routers too.

- Add ExternalExpressRoutes as a registry of external routes
- Add ExternalRoute as the route entry returned by `findRoute` action
- Move implementation details of StaticAssetsRoute to these new classes
- Remove StaticAssetsRoute
